Break up OpenModal into smaller components
