+++ src/core/components/controls/OpenControl.tsx
@@ -3,7 +3,7 @@ import { observer } from 'mobx-react';
 import BaseControl from './BaseControl';
 import Modal from 'react-modal';
 import modalStyle from '@data-story-org/core/src/utils/modalStyle';
-import OpenModal from '../modals/OpenModal';
+import OpenModal from '../modals/Open';
 import { Store } from '../../store';
 
 interface Props {
