+++ src/core/components/modals/Open/Open.tsx
@@ -1,0 +1,69 @@
+import React, { FC, useState } from 'react';
+import { observer } from 'mobx-react';
+import Cookie from '../../../utils/Cookie';
+import DiagramModel from '../../../DiagramModel';
+import { Store } from '../../../store';
+import OpenModalHeader from './OpenHeader';
+import OpenModalBody from './OpenBody';
+import OpenModalActions from './OpenActions';
+
+interface Props {
+  store: Store;
+  closeModal: () => void;
+}
+
+const OpenModal: FC<Props> = ({ store, closeModal }) => {
+  // Never used
+  /* const [storyName, setStoryName] = useState(''); */
+
+  /* const handleChange = (e) => {
+   *   setStoryName(e.target.value);
+   * }; */
+
+  const handleCancel = (_e) => {
+    closeModal();
+  };
+
+  const handleClear = (_e) => {
+    Cookie.clear();
+    store.setStories(Cookie.keys());
+  };
+
+  /* const handleSave = (e) => {
+   *   //
+   * }; */
+
+  const clickStory = (name) => {
+    try {
+      const engine = store.diagram.engine;
+      const model = new DiagramModel();
+      model.deserializeModel(
+        Cookie.getObject(name),
+        engine,
+      );
+      engine.setModel(model);
+      closeModal();
+    } catch (e) {
+      alert(
+        `Could not create engine for story ${name}. See console for details.`,
+      );
+      console.error(e);
+    }
+  };
+
+  return (
+    <div>
+      <OpenModalHeader />
+      <OpenModalBody
+        store={store}
+        clickStory={clickStory}
+      />
+      <OpenModalActions
+        handleClear={handleClear}
+        handleCancel={handleCancel}
+      />
+    </div>
+  );
+};
+
+export default observer(OpenModal);
