+++ src/core/components/modals/Open/index.ts
@@ -1,0 +1,3 @@
+import OpenModal from './Open';
+
+export default OpenModal;
